+++ src/main/java/com/example/charging_life/station/StationController.java
@@ -26,10 +26,10 @@ import java.util.List;
 @RestController
 @RequiredArgsConstructor
 
-@Tag(name = "ChargingStaion API" , description = "<공공 api 받아오는 API>" + "\n\n" +
-        "⏰ 매주 금요일 5:00 AM 마다 새로운 충전소의 공공 api를 Business 데이터베이스에 저장 " + "\n\n" +
-        "⏰ 매주 금요일 5:30 AM 마다 새로운 충전소의 공공 api를 ChargingStation & Charger 데이터베이스에 저장" + "\n\n" +
-        "⏰ 5분마다 수행하여 성공하면 공공 api를 갱신하여 Charger 데이터베이스에 update")
+@Tag(name = "ChargingStaion API" , description = "<충전소 및 충전기 정보 조회 API>" + "\n\n" +
+        "\uD83D\uDCCC 충전소 정보와 즐겨찾기 여부 및 몇명이 접근하고있는지에 대한 정보 반환 " + "\n\n" +
+        "\uD83D\uDCCC 충전소 이름(statNm) 또는 광역시,도명(city)을 이용해 충전소를 검색" + "\n\n" +
+        "\uD83D\uDCCC 관리자가 등록한 충전소 기준으로 조회")
 public class StationController {
     @Value("${key.serviceKey}")
     private String key;
@@ -39,36 +39,6 @@ public class StationController {
     private final PasswordEncoder passwordEncoder;
     private final StationService stationService;
 
-
-    @Operation(summary = "공공 api 받아오기", description = "성공하면 공공 api를 ChargingStation & Charger 데이터베이스에 저장")
-    @GetMapping("/station/api")
-    public void getChargingStationApi() throws IOException, ParseException {
-        stationService.saveChargingStationData(false,false);
-    }
-
-    @Operation(summary = "새로운 공공 api 추가하기", description = " 매주 금요일 5:00 AM 마다 수행하여 성공하면 새로운 충전소의 공공 api를 Business 데이터베이스에 저장")
-    @Scheduled(cron = "0 00 05 * 6 *")
-    public void updateBusinessApi() throws IOException, ParseException {
-        stationService.saveChargingStationData(true,true);
-    }
-
-    @Operation(summary = "새로운 공공 api 추가하기", description = " 매주 금요일 5:30 AM 마다 수행하여 성공하면 새로운 충전소의 공공 api를 ChargingStation & Charger 데이터베이스에 저장")
-    @Scheduled(cron = "0 30 05 * 6 *")
-    public void updateChargingStationApi() throws IOException, ParseException {
-        stationService.saveChargingStationData(false,true);
-    }
-
-//    @Operation(summary = "5분마다 공공 api 갱신", description = " 5분마다 수행하여 성공하면 공공 api를 갱신하여 Charger 데이터베이스에 update")
-//    @Scheduled(fixedDelay = 18000000)
-//    public void updateCharger() throws IOException, ParseException {
-//        stationService.updateChargerData();
-//    }
-
-    @Operation(summary = "공공 api 받아오기", description = "성공하면 공공 api를 Business 데이터베이스에 저장")
-    @GetMapping("/business/api")
-    public void getBusinessApi() throws IOException, ParseException {
-        stationService.saveChargingStationData(true,false);
-    }
 
     @Operation(summary = "해당 statId 충전소 정보",
             description = "충전소 정보와 즐겨찾기 여부 및 몇명이 접근하고있는지에 대한 정보 반환 ")
